+++ src/hsoines/oekoflex/bid/PowerBid.java
@@ -8,7 +8,7 @@ import hsoines.oekoflex.energytrader.MarketOperatorListener;
  * Time: 21:34
  */
 public abstract class PowerBid extends BidSupport {
-    public PowerBid(final float price, final int quantity, final MarketOperatorListener marketOperatorListener) {
+    public PowerBid(final float price, final float quantity, final MarketOperatorListener marketOperatorListener) {
         super(price, quantity, marketOperatorListener);
     }
 }
