+++ src/hsoines/oekoflex/bid/PowerPositive.java
@@ -10,7 +10,7 @@ import hsoines.oekoflex.energytrader.MarketOperatorListener;
  */
 public final class PowerPositive extends PowerBid {
 
-    public PowerPositive(float price, int quantity, MarketOperatorListener marketOperatorListener) {
+    public PowerPositive(float price, float quantity, MarketOperatorListener marketOperatorListener) {
         super(price, quantity, marketOperatorListener);
     }
